Removing some more target / platform specific code out of main.c and simplifying initialisation of system
