+++ src/main/drivers/system_stm32f10x.c
@@ -15,6 +15,7 @@
  * along with Cleanflight.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+#include <string.h>
 #include <stdbool.h>
 #include <stdint.h>
 #include <stdlib.h>
@@ -22,9 +23,13 @@
 #include "platform.h"
 
 #include "gpio.h"
+#include "nvic.h"
 #include "system.h"
 
 #define AIRCR_VECTKEY_MASK    ((uint32_t)0x05FA0000)
+
+// from system_stm32f10x.c
+void SetSysClock(bool overclock);
 
 void systemReset(void)
 {
@@ -39,7 +44,6 @@ void systemResetToBootloader(void) {
     *((uint32_t *)0x20004FF0) = 0xDEADBEEF; // 20KB STM32F103
     systemReset();
 }
-
 
 void enableGPIOPowerUsageAndNoiseReductions(void)
 {
@@ -61,3 +65,48 @@ bool isMPUSoftReset(void)
     else
         return false;
 }
+
+void systemInit(void)
+{
+    SetSysClock(false);
+    
+#ifdef CC3D
+    /* Accounts for OP Bootloader, set the Vector Table base address as specified in .ld file */
+    extern void *isr_vector_table_base;
+
+    NVIC_SetVectorTable((uint32_t)&isr_vector_table_base, 0x0);
+#endif
+    // Configure NVIC preempt/priority groups
+    NVIC_PriorityGroupConfig(NVIC_PRIORITY_GROUPING);
+
+    // Turn on clocks for stuff we use
+    RCC_APB2PeriphClockCmd(RCC_APB2Periph_AFIO, ENABLE);
+
+    // cache RCC->CSR value to use it in isMPUSoftreset() and others
+    cachedRccCsrValue = RCC->CSR;
+    RCC_ClearFlag();
+
+    enableGPIOPowerUsageAndNoiseReductions();
+
+    // Set USART1 TX (PA9) to output and high state to prevent a rs232 break condition on reset.
+    // See issue https://github.com/cleanflight/cleanflight/issues/1433
+    gpio_config_t gpio;
+
+    gpio.mode = Mode_Out_PP;
+    gpio.speed = Speed_2MHz;
+    gpio.pin = Pin_9;
+    digitalHi(GPIOA, gpio.pin);
+    gpioInit(GPIOA, &gpio);
+
+    // Turn off JTAG port 'cause we're using the GPIO for leds
+#define AFIO_MAPR_SWJ_CFG_NO_JTAG_SW            (0x2 << 24)
+    AFIO->MAPR |= AFIO_MAPR_SWJ_CFG_NO_JTAG_SW;
+
+    // Init cycle counter
+    cycleCounterInit();
+
+    memset(extiHandlerConfigs, 0x00, sizeof(extiHandlerConfigs));
+    // SysTick
+    SysTick_Config(SystemCoreClock / 1000);
+}
+
