+++ src/main/drivers/system_stm32f30x.c
@@ -15,6 +15,7 @@
  * along with Cleanflight.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+#include <string.h>
 #include <stdbool.h>
 #include <stdint.h>
 #include <stdlib.h>
@@ -22,9 +23,11 @@
 #include "platform.h"
 
 #include "gpio.h"
+#include "nvic.h"
 #include "system.h"
 
 #define AIRCR_VECTKEY_MASK    ((uint32_t)0x05FA0000)
+void SetSysClock();
 
 void systemReset(void)
 {
@@ -75,4 +78,27 @@ bool isMPUSoftReset(void)
         return true;
     else
         return false;
+}
+
+void systemInit(void)
+{
+    // Enable FPU
+    SCB->CPACR = (0x3 << (10 * 2)) | (0x3 << (11 * 2));
+    SetSysClock();
+
+    // Configure NVIC preempt/priority groups
+    NVIC_PriorityGroupConfig(NVIC_PRIORITY_GROUPING);
+
+    // cache RCC->CSR value to use it in isMPUSoftreset() and others
+    cachedRccCsrValue = RCC->CSR;
+    RCC_ClearFlag();
+
+    enableGPIOPowerUsageAndNoiseReductions();
+
+    // Init cycle counter
+    cycleCounterInit();
+
+    memset(extiHandlerConfigs, 0x00, sizeof(extiHandlerConfigs));
+    // SysTick
+    SysTick_Config(SystemCoreClock / 1000);
 }
